requirements.txt
albumentations==0.3.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.5
keras-rectified-adam==0.9.0
numpy==1.15.4
opencv-python==3.4.1.15
opencv-python-headless==4.1.0.25
pandas==0.24.2
pandocfilters==1.4.2
Pillow==6.1.0
pretrainedmodels==0.7.4
scikit-image==0.14.1
scipy==1.1.0
tqdm==4.28.1
